+++ backend/src/main/java/com/cerebrus/tema/TemaService.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public interface TemaService {
 
-    List<Tema> ObtenerTemasPorCursoAlumno(Integer cursoId);
+    List<Tema> ObtenerTemasPorCursoAlumno(Long cursoId);
 
-    List<Tema> ObtenerTemasPorCursoMaestro(Integer cursoId);
+    List<Tema> ObtenerTemasPorCursoMaestro(Long cursoId);
 
     Tema crearTema(String titulo, Long cursoId, Long maestroId);
